requirements.txt
# requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
Pillow==9.5.0
celery==5.2.7
redis==4.5.4
boto3==1.26.118
python-dotenv==1.0.0
gunicorn==20.1.0
whitenoise==6.4.0
django-storages==1.13.2
drf-yasg==1.21.5  # For API documentation